Close accepted privilege escalation; order governance checks by jurisdiction
Both fixes were driven by pre-existing failures in Phase28StressTests and
Phase29StressTests. No test expectation was relaxed.

ContractAcceptanceEngine took value, norm and role stores but no ContractStore,
and evaluate() received only the proposed contract, so it judged every proposal
in isolation and could not see escalation relative to grants already held. The
Temporal Privilege Creep scenario demonstrated it: a day-1 contract scoped
speak_only forbidding execute/navigate, followed by a day-20 contract scoped
full_access with no forbidden actions, was ACCEPTED.

Adds an optional ContractStore constructor argument and Check 8, catching
escalation by omission (dropping a restriction an active contract imposed) and
scope broadening (a specific grant widening to full_access/all/unrestricted).
Response is MODIFIED rather than REJECTED: scope is narrowed and inherited
restrictions are carried forward, so a later contract cannot acquire authority
by silence. The 3-argument constructor is unchanged, so existing call sites keep
the previous isolated-evaluation behaviour.

Both engines also evaluated the universal ABSOLUTE-value floor early, so it
short-circuited narrower checks and those paths were never reached:

  ContractAcceptanceEngine  role scope now precedes ABSOLUTE values
  RoleAcceptanceEngine      provenance -> role content -> ABSOLUTE values

The SET of rejected contracts and roles is unchanged - only which rule is
reported as governing - because a proposal tripping only the value check still
reaches it.

Also fixes a test that exited 0 while printing "Some tests FAILED" (1 of 128
assertions was failing silently and would have read as green in CI), and a
missing <iomanip> that GCC 16 no longer provides transitively via <iostream>.

Phase 28: 6/8 -> 8/8.  Phase 29: 16/20 -> 20/20.  Full sweep: 37 build, 34 pass,
no regressions.

Notes:
- include/contracts/ and include/roles/ were untracked; the governance engines
  enter version control with this commit.
- test_developmental_tracking.cpp carries 4 pre-existing static_cast narrowing
  fixes that were already in the working tree.
- The learning-rate default fix in src/main.cpp is NOT included: that file has
  ~6,600 lines of unrelated uncommitted work and committing it would sweep that
  up. It remains in the working tree.
- Landed without governance re-ratification by owner decision; see
  Validation_notes/governance_engine_changes_2026-08-22.md

+++ src/test_developmental_tracking.cpp
@@ -1,5 +1,6 @@
 #include "core/LanguageSystem.h"
 #include <iostream>
+#include <iomanip>  // std::setprecision (not transitively provided by <iostream> in GCC 16)
 #include <vector>
 #include <chrono>
 #include <random>
@@ -129,7 +130,7 @@ private:
         // Caregiver-like interactions with motherese
         if (step % 8 == 0) {
             std::vector<std::string> babbling_sounds = {"mama", "baba", "dada", "gaga", "nana"};
-            std::uniform_int_distribution<> sound_dist(0, babbling_sounds.size() - 1);
+            std::uniform_int_distribution<> sound_dist(0, static_cast<int>(babbling_sounds.size()) - 1);
             
             // Higher pitch for motherese
             auto teacher_audio = generateSyntheticAudio(babbling_sounds[sound_dist(rng_)], 200.0f, true);
@@ -145,7 +146,7 @@ private:
                 {"ball", {0.2f, 0.3f}},
                 {"cup",  {0.7f, 0.4f}}
             };
-            std::uniform_int_distribution<> obj_dist(0, early_objects.size() - 1);
+            std::uniform_int_distribution<> obj_dist(0, static_cast<int>(early_objects.size()) - 1);
             auto [word, gaze] = early_objects[obj_dist(rng_)];
             language_system_->processJointAttentionEvent(gaze, word);
         }
@@ -182,7 +183,7 @@ private:
                 {"toy", {0.1f, 0.8f}}
             };
             
-            std::uniform_int_distribution<> word_dist(0, grounded_words.size() - 1);
+            std::uniform_int_distribution<> word_dist(0, static_cast<int>(grounded_words.size()) - 1);
             auto [word, gaze] = grounded_words[word_dist(rng_)];
             
             // Teach word with visual grounding
@@ -199,7 +200,7 @@ private:
         // Advanced communication patterns
         if (step % 5 == 0) {
             std::vector<std::string> advanced_words = {"please", "thank", "help", "want", "like"};
-            std::uniform_int_distribution<> word_dist(0, advanced_words.size() - 1);
+            std::uniform_int_distribution<> word_dist(0, static_cast<int>(advanced_words.size()) - 1);
             
             std::string word = advanced_words[word_dist(rng_)];
             auto teacher_audio = generateSyntheticAudio(word, 170.0f);
